feat: rename Achievement to AchievementEntity and enhance attributes
